requirements.txt
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Pillow==9.2.0